move channel tests to a separate file
